Creates a new RStudio project with the exercises and needed data sets

Creates a new RStudio project with the exercises and needed data sets

bdc_init_class_local(
  folder = "big-data-class",
  avg_daily_orders = 100,
  avg_no_items = 3,
  days_in_segment = 10,
  no_of_segments = 100,
  seed = 7878,
  product_data = bdc_create_products(),
  customer_data = bigdataclass::customers,
  start_date = "2016-01-01",
  no_transactions = 50000,
  batch_size = 5000,
  no_files = 5,
  src = system.file("exercises", package = "bigdataclass"),
  overwrite = FALSE,
  include_books = FALSE
)

Arguments

folder
Path to folder where to build the project

avg_daily_orders
Number of average daily orders

avg_no_items
Average number of sales in an order

days_in_segment
Days in a segment

no_of_segments
Number of segments

seed
Seed to randomize to

product_data
Product data frame source

customer_data
Customer data frame source

start_date
The start date for the orders

no_transactions
Number of transactions to save to file

batch_size
File batch size

no_files
Number of files

src
Path to the source folder of the exercises

overwrite
Overwrite class folder if it exists

include_books
Download and save the book data